+++ default.nix
@@ -1,0 +1,17 @@
+let
+  holonixPath = (import ./nix/sources.nix).holonix;
+  holonix = import (holonixPath) {
+    holochainVersionId = "v0_0_165";
+    include = {
+      niv = false;
+      scaffolding = false;
+      happs = false;
+      git = false;
+    };
+  };
+  nixpkgs = holonix.pkgs;
+in
+nixpkgs.mkShell {
+  inputsFrom = [ holonix.main ];
+  packages = with nixpkgs; [];
+}